fix review problems
