+++ server/src/main/java/org/apache/iotdb/db/qp/physical/crud/AggregationPlan.java
@@ -176,19 +176,18 @@ public class AggregationPlan extends RawDataQueryPlan {
 
   @Override
   public String getColumnForDisplay(String columnForReader, int pathIndex) {
-    String columnForDisplay = columnForReader;
     if (isGroupByLevel()) {
-      if (resultColumns.get(pathIndex).hasAlias()) {
-        return resultColumns.get(pathIndex).getAlias();
-      }
-
       PartialPath path = paths.get(pathIndex);
       String functionName = aggregations.get(pathIndex);
       String aggregatePath =
           groupByLevelController.getGroupedPath(
               String.format("%s(%s)", functionName, path.getFullPath()));
-      columnForDisplay = aggregatePath;
+      String alias = groupByLevelController.getAlias(aggregatePath);
+      if (alias != null && !"".equals(alias)) {
+        return alias;
+      }
+      return aggregatePath;
     }
-    return columnForDisplay;
+    return columnForReader;
   }
 }
