+++ server/src/main/java/org/apache/iotdb/db/query/dataset/AlignByDeviceDataSet.java
@@ -58,7 +58,7 @@ public class AlignByDeviceDataSet extends QueryDataSet {
   private QueryContext context;
   private IExpression expression;
 
-  // contains original measurement, no include alias
+  // contains only original measurement, alias not included
   private List<String> measurements;
   private List<PartialPath> paths;
   private List<String> aggregations;
