+++ server/src/main/java/org/apache/iotdb/db/utils/QueryDataSetUtils.java
@@ -67,8 +67,7 @@ public class QueryDataSetUtils {
       if (queryDataSet.hasNext()) {
         RowRecord rowRecord = queryDataSet.next();
         // filter rows whose columns are null according to the rule
-        if (queryDataSet.withoutNullFilter(
-            rowRecord, queryDataSet.isWithoutAllNull(), queryDataSet.isWithoutAnyNull())) {
+        if (queryDataSet.withoutNullFilter(rowRecord)) {
           // if the current RowRecord doesn't satisfy, we should also decrease
           // AlreadyReturnedRowNum
           queryDataSet.decreaseAlreadyReturnedRowNum();
